simulator: add offset in reported object positions
This also happens with the vision, though slightly different.
Here, the offset is constant over the whole simulated camera, but in
reality, this offset will vary over the area of the camera, resulting
from wrong distortion settings or calculation.
During the robocup 2019 (towards the final), the difference in reported
positions of the two cameras in their overlap area was relatively
constant for most of the overlap area, so this simplified model was
chosen to simulate this effect.
It should re-create most of the annoying artefacts this can cause.
The value of 2 cm per camera (a total difference of 4 cm difference
between the two cameras) was chosen as it is close to the values
observed during robocup 2019.

Should a filter be implemented in the tracking to combat this effect,
real vision data from games has to be used, as the effect simulated here
is highly simplified.