amun: simulator: do not produce an offset smaller than 1 nm
This is mostly because 0 < 0 <=> false, and so the "just use the camera instead of normalizing" part did not work.
Additionally, very small offsets just do not make sense, get lost in everything from noise to calculation issues.
In order to avoid situations where e ist just big enough not to be 0, but e**2 is not,
I introduced this limit to 1 nm
which should be totally ok