Glados: o/modificationchecker: also check for NaN
Without these added lines, the modificationchecker already threw and error if any of these values are NaN.
This is because even if they are not modified, NaN != NaN is true and so the old and the new value would be
not equal. However, the old message was highly confusing. A robot / ball was modified during the strategy execution,
even if we did not change anything in the stragey but in ra?

The new dedicated checks allow us to add a more useful error message when we start executing
the strategy with NaN as speeds or positions.