cmake: sourcemap.patch: Use std::bind correctly
The old code would have resulted in a template specialization failure.
However, as these calles are themselves under a template that will not be specialized by
our framework right now, the effect was not noticable.

The correct usage of std::bind has to specify ALL parameters to the given functor,
but can use placeholders to avoid specializing all arguments