amun: trajectorypath: use direct trajectory more often
The direct trajectory was always directly discarded when it got too
close to an obstacle, and the standard sampler was used.
However, even if it is too close, it might still be the best trajectory
as judged by the standard sampler.
Therefore, the direct trajectory is also evaluated by the criteria of
the standard sampler and used if it is better.

This can result in better paths and most importantly less change in the
used path if an obstacle comes close to the direct trajectory.
This activates most often when the start point is already close to an
obstacle.