glados:hwchallenges: correctly handle ball near obstacles
If the vision quality of the ball drops to low we assume that we don't
know where the ball is. This is to avoid the robot moving/turning
rapidly when it's near the ball.
Add Obstacle for the ball if the robot can't put it into its dribbler to
avoid it pushing the ball away. E.g if the robot moves backwards towards
the ball instead of around it.