tracking: ballgroundcollisionfilter: fix ball not staying put after being pushed
When the robot pushes the ball (or dribbles it forwards) and it is
invisible, the robot stops and drives back,
the intended behavior of the tracking is that the ball stays where it is
pushed to so long as that position stays invisible.
Once the position the ball was pushed to becomes visible and the ball is
not there, the tracked ball should snap back to the pushing robot.

This has an additional check where if the position that the ball was
pushed to is shadowed by another robot, then it is assumed to stay
attached to the dribbling robot (this is mostly used for duel scenarios
and similar).

This check caused the whole thing to malfunction here, since the pushing
robot was always used for the shadowing check of the other robots.