tracking: ballgroundcollisionfilter: add hysteresis to invisible ball check
For pushing and pulling the ball this is not really an issue since the
robot will be travelling rather fast and in one direction.
But this code also activates for rotating with the ball while dribbling,
in this case the ball can be invisible and the last known ball position
can be just on the border of being visible and invisible.
In this case a hysteresis will be necessary.