tracking: ballgroundcollisionfilter: fix ball velocity while dribbling forward
.. with an invisible ball.
While dribbling forward, the velocity was previously set to zero since
both pushingPosVisible and otherRobotObstructions were of course false
and it went into the invisible standing ball branch.
This is not much of an issue, except that the ball velocity is set to
zero instead of the robot velocity.
Using the dribbling if-branch while the ball is being pushed forward
solves this issue.