tracking: ballgroundcollisionfilter: fix chaseball condition
The previous commit fixed a bug which now activated this condition more
often. This now lead to the case that the chaseball condition activated
too often, leading to problems.
This is the case when the ball is shot at the robot and the projected
futureball is already at the other side of the robot and the current
position is slightly inside the robot.
Then, only the ouside robot projection activated and the ball was
projected to the backside of the robot.