tracking: ballgroundcollisionfilter: fix using direction instead of position for intersection
Interestingly, this bug did not cause big problems since the factor 1000
is rather large.
This is because lim x->inf (pastPos + projectDir * x).normalized() ==
projectDir.normalized() holds true.

But of course the function intersectLineSegmentRobot takes two positions
and not a position and a direction, so the new version is correct.