tracking: ballgroundcollisionfilter: fix intersectLineSegmentRobot function
The bug was when p1 is still inside the circle of the robot but in front
of the dribbler (thus not in the robot) and p2 is far in front of the
dribbler outside the circle.
The previous code then returned a position on the circle in front of the
dribbler as the intersection, which is not correct.