tracking: ballgroundcollisionfilter: rework ball robot intersection logic
The main point here is to make writeBallState mostly non-modifying so
that it can be called with different times (even going backwards in
time) without problems.
This is only possible if the intersection logic is handled
independently, in this case it happens when the ball is not visible for
a frame (although the call is still in writeBallState, it is now
conditional and depends on whether there was an empty vision frame
before, this is equivalent to calling updateEmptyFrame from the outside
when necessary).

Some of the funcitonaly (rotating, volley shots) are not yet enabled
again since they too need to be modified to work with this new paradigm.