tracking: ballgroundcollisionfilter: better handling of slowly driving up to the ball
See attached unit test for a situation where this happens.
Due to partial ball occlusion, the ball moves a bit when the robot is
driving up to it, which often creates problems. Additionally, the ball
might not intersect with the robot even though it properly touches the
dribbler.
This code activates dribbling for these cases.