seshat: combinedlogwriter: insert status with vision geometry at beginning of log
This is to ensure that creating small logs for tracking replays works
properly since the ssl vision only sends out the geometry roughly every
500ms.
If the log is loaded and tracking replay is enabled without having
played anything, the first status package in the log will NOT reach the
tracking replay since it is automatically played when opening the log.
Therefore, this information must be put in the second frame of the log
and not the first.