amun: processor/trackingreplay: use sides is flipped from the log
In order to always have the ball at the same position in the log and in
the replay.
This is done by just using the whole referee information present in the
log, not just the sides is flipped information. This is more general and
should be less prone to errors.