tracking: ballflyfilter: also abort potential chips due to collisions
m_isActive is only true when the flight filter also writes out its
results instead of the ground filter.
However, a chip can be detected and tracked while m_isActive is false if
the reconstruction is not good enough.
This frequently happens during linear kicks.
In these cases, it is important to also abort the tracking after a
collision of the ball with the robot since the unusual trajectory could
otherwise result in an errouneous activation of the filter.
Also, the robot that the ball collided with could have initiated a true
chip kick, which would then not be properly detected.