amun: seshat: fix tracking replay not getting a team status from loglogs made from backlogs
The problem is twofold.
The replay logger in seshat only receives status packages from the
backlog in this scenario, and there likely is not a single team update
status in there.
Therefore, when opening a backlog, add the team information from the
regular logger to the replay logger.

The second issue is that when the tracking replay is performed in ra,
the first frame of the log containing the team information is preloaded
before the user gets the chance to activate tracking replay, meaning
that the information is lost to the tracking replay.
This is solved by injecting an empty status package in front of the team
information.