tracking: ballflyfilter: consistently use float for time differences
Since ChipDetection.time is a time difference to the initialization of
the frame, using a float is more than enough (and when the value is
stored in the struct, it is already a float).
This gets rid of inconsistent use of float and double which can be
disorienting and reduces performance.