tracking: ballflyfilter: use pure capture time for flight reconstruction
The tracking usually translates the capture time to one that fits the
local computer time. This is because the time on the vision computer and
the own computer might differ by a few seconds, which can create
substantial problems.

The flight reconstruction only needs the relative times, so using the
raw capture time is fine here.
This is done since the translation of the time (sourceTime in
tracker.cpp) adds additional noise in the form of network and threading
delays.
The problem is that the flight reconstruction with least squares fitting
is very sensitive to incorrect times by its nature.
Therefore, eliminating this additional noise significantly improves the
reconstruction accuracy.