tracking: ballflyfilter: move bounce updating away from predictTrajectory
predictTrajectory (most importantly used in writeBallState) should NEVER
update the internal state. writeBallState can be called with times going
back and forth and is also far ahead of the time in the vision
detections.

In this case, this lead the m_bouncing activating a good while before
the ball actually bounced, and since this change can not be taken back,
further updates to the trajectory where not possible.

predictTrajectory could also not be used for chooseDetection because of
the missing control over the input time.