tracking: ballflyfilter: remove the area based chip detection and reconstruction
It does not properly work anymore with the current ssl camera setup.
While the ball area obviously still becomes larger when the ball is
higher up, with the high mounting point of the cameras, this effect is
much smaller than before.
Also there is high noise present in the area data, which makes using it
effectively hard.
In general, it might be possible to use this data later on in the flight
once enough data is collected to smooth out the noise, but at that point
the regular flight fitting will always produce better results anyways.
Using the area approach to reconstruct the flight in its early phases is
not reasonably possible.