tracking: ballflyfilter: do not use linear/chip command if both were recently given
Previously, only the command of the first frame of the shot was used to
determine linear or chip if the information is present.
This could be an issue when the strategy changed this during/after the
shot since the tracking can not accurately determine in which frame the
ball was actually shot.