tracking: ballgroundcollisionfilter: fix accepting detection far off during chip shots
The problem is that m_lastReportedBallpos is only updated during
writeBallState, which is not called if the flight tracker is active.
Therefore, if if there is a chip shot ongoing and a stray ball detection
is found near the start point of the shot, the ground filter will accept
the detection and go haywire.