tracking: track robots and ball with the same camera when possible
It would have been possible to keep only one filter per robot and just
change the primary camera for this filter.
However, given that the ground collision filter needs the robot
positions, it is best if multiple filters are used since the positions
for the ball shadowing computation will not match otherwise.