tracking: ballflyfilter: add detection capture start time to reconstructed flight
This is currently only used to improve the accuracy of chipShotError.
The issue is that flight reconstruction is optimized for the capture
time, so the error calculation must also use the capture time to be
accurate.

The issue was that the condition linearShotError(...) >
chipShotError(...) was no longer true for a short while in the flight in
many cases, which created multiple flight activations.