amuncli: save backlogs when a certain command string is found in the log
Similar to the handling of os.exit(1), amuncli.requestBacklog(TEST) can
be printed by any part of the software (for example the autoref,
strategy, ...) to save a backlog.
For this to work, saving backlogs must be enabled in the amuncli.
Also take care that the string in the requestBacklog "function call" is
a valid directory name, i.e. it must no contain weird characters.