amun: trajectorypath: use constant time intervals for endinobstacle obstacle checking
Consider the case where the robot will, due to its current velocity,
inevitably run into an obstacle, but just by a little.
The endinobstacle sampler could sometimes still find a "valid" path if
the target position of that path was far enough away.
Since for such a long trajectory the total time is very large, the time
intervals for obstacle checking are also very large and can possibly be
aligned such that none of the sampled points on the trajectory actually
intersect with the obstacle.

This is fixed by simply setting a constant sampling time interval.
The selected interval is slightly smaller than the previoyusly average
interval size.