glados: a/a/freekick: Use shared util shootGoalPossible
It's useful to have this centralized, since the logic is shared with
a/a/shoot

Omitting the `RefereeState === "DirectOffensive"` check is okay, since
a/a/freekick runs only during free kicks anyways.

Checking whether World.Ball.pos.y > -0.2 was introduced by 3f9838567bb9.
The original intention was to *not* try goal shots with the ball on the
friendly half, however the conditional is inverted. This was okay
though, since it ensured we try shooting at the goal during kickoff.
It is important we are able to score goals in cases where the opponent
has no robots on the field or is not able to drive after kickoff. Are
more refined version should use observer/crash and interact with the
dirty flag (also see #921), for now, this is fine though.