glados: t/at/placeball: fix PULL_TO_FIELD
Previously, when the ball was outside the field, but almost perfectly
aligned with a field line (see log), we pulled the ball back towards the
field, and it landed just outside the field, on the field line. (This
could also happen when the ball moves sideways while dribbling.)

In this situation, the previous code tried to pull the ball back again
(as is was still outside the field), but the border offset (and so the
direction of the pull) wasn't recalculated because the ball was already
so close to were it was supposed to be. So we pull the ball back in
the same direction and it lands just outside the field again.

The fix is to limit the ball position not just to the field, but to the
field minus the robot and ball radius. That way there is no chance that
the ball is outside the field after the pull.