git/gitlive: use create diffopts differently
Also moves up minimum required gcc version.

gcc 7.5.0 (used in our CI) is notoriously fiddely when working with
nontrivial designated initializers. As this gitlive.cpp is the adaptor
between the c library libgit2 and our framework, it uses designated
initializers to make the code a bit more readable.

However, as 7.5 seems to be confused after .pathspec (a nested
designated initializer), and is not happy with reodering either,
this code now omits everything after the nested designated initializers,
and sets them afterwards manually. A good compiler is going to optimize
that away anyways.