glados:objective: improve midfield zones and refactoring
Replace counter that counts upwards with a maximum to a counter that
counts downwards until zero to make the API more intuitive.
The zone splitting code now takes SplitZone[] instead of Boundaries[], which
allows for more control for the caller.
The midfield zones now contain an offensive zone that gets split at most
one time.