glados: t/a/shoot: do not use chaseball for low flying balls hitting the robot
The issue was with a ball that is still slightly jumping.
The chaseball future ball was on the other side of the robot compared to
the current ball.
The condition that checks that the chase ball pos is on the same side of
the robot as the current ball pos made an exception if the ball is
currently flying.
This is to allow using chaseball when a large chip shot goes over the
robot. This leads to a problem when the ball hits the ground before
arriving at the chaseball future pos, which is fixed by this commit.