glados: u/defense: Rework dangerous rating using time estimation
Replace the previous heuristic for how dangerous a situation is, by an
approximation for how fast an opponent robot could possibly shoot a
goal given the current ball position.
Looking at logs and simulation this seems to be a more helpful idea and
has the additional effect that it flickers less when the ball moves fast.