glados: c/messaging: Implement impersonation
To avoid messaging delays, the trainer must be able to send an objective
when selecting a new main attacker. A proper solution would specify the
sender type of selectedObjective to be "trainer" | "robot". However, the
messaging type system does not handle this case at the moment (i.e. when
receiving through the normal receive functions instead of through
receiveTrainer, the sender is always assumed to be a FriendlyRobot).
For this reason, we allow the trainer to impersonate other robots.

See #936