glados: tr/roles: Send out an objective when selecting the main attacker
This fixes a bug where the number of attackers that are part of the
support group flickers for two frames on a main attacker switch.

Because of messaging delay, the support group makes an assumption: the
number of of attackers that are part of the support group always stays
the same on a main attacker switch. Prior to the introduction of
objectives, the procedure was the following:

Frame 0
    R1
        is MA
        is not member
    R2
        is member
        send group application
Frame 1
    R1
        is not member
        send group application
    R2
        is MA
        is member
Frame 2
    R1
        is member
        send group application
    R2
        is MA
        is not member

The number of members stays constant, the number of zones does not
change. With the introduction of objectives, the MA selects an
objective during its first active frame (frame 1). The objective first reaches
the supports in frame 2, thus, R1 only starts sending out its group
application in frame 2 and becomes a member of the support group in
frame 3.

Because of this, there is a single frame (frame 2), where neither R1 nor R2 are
member of the support group, leading to zone flickering.

The solution is to allow the trainer to send out an objective when
selecting a new main attacker. After the trainer is run, messages are
delivered. The support attackers thus receive the objective in frame 1
and send out a group application.

Pieces of code that previously sent a predefined objective (defense
code) now simply propagate the trainer objective instead of sending
their own choice. In these cases, we may want advanced objective
selection in the future but for now, propagating the trainer objective
instead of a static one should be better.

The trainer also gains the capability to cancel messages: the old main
attacker sent out an objective but there may only be a single active
objective at a time. The old main attacker does not know, whether it
will keep on being main attacker. Only the trainer has this knowledge,
thus it has to cancel the message.