glados: use the advantage rule in some cases
The advantage rule has been reintroduced, but differently.
This time if the opposing team receives a yellow card, but does not
remove a robot in 10s the game would be halted, unless we send an
advantage rule. Some of the code from this commit has been "stolen" from
code that was removed in 9f2d6fc1a4f4022da96d170eac526dcc14c26776, but
the heuristics and the triggering event for the advantage changed,
because of the above change in the advantage rule.