glados: a/b/behavior: Allow behaviors to keep their current task running without supplying constructor arguments
It is possible for the constructor arguments to a task to be only
available when first starting the task (for example: a task is started
after a pass is played and requires the pass info message, however that
message will not be available later). Instead of caching them on the
behavior, I think its preferrable to allow continued use of the old
task.