glados: u/volley: Fix crash
calculateVOutTeamCoordinates no longer crashes when the veloclity given
in v_out_length is physically not possible. Instead undefined is
returned which needs to be handled by the caller. Adds a function that
calculates the volley direction with minimal possibility. This is used
as a fallback for observer/goal since the slowest direction is the most
dangerous. This is because faster shots are more likely to be detected
by the autoref.