glados: g/m/armada: Fix assignment bug
In armada the hungarian algorithm is used to assign 4 robots to 4
positions. However, 5 robots are participating in this move. The robot
with index 0 is assigned the freekick task and therefore doesn't need a
position assignment. That means we want to assign robots with indices 1,
2, 3, 4 to positions with indices 0, 1, 2, 3. Since the implementation
of the hungarian algorithm just returns a permutation of the numbers 0,
1, 2, 3, we need to offset them by +1 when indexing this._robots.