glados: a/b/objective: Retain objective instance across MA switches
Some supporters may keep state in a behavior. This state is lost when
the objective is reconstructed, thus we keep the objective alive.

When an agent becomes main attacker, we delete its old support runner. We
could also call stop on the old instance, but *should* be no real
difference, as stop *should* put the behavior in the same state as its
construction.