glados: tr/trajectorypath: rotate slower when target is still far away
When the robot is fully accelerating and tries to rotate at the same
time, the maximum acceleration of the motors is not sufficient.
This can result in the trajectory not being correctly followed.
Therefore, the acceleration and maximum velocity for the rotation is
reduced until the target position is close.