ra: mainwindow: minimize the number of strategy reloads when starting ra
The problem was that the strategy was first loaded and then the geometry
for the simulator etc. was loaded.
This lead to reloading the strategy multiple times.
The first part of the solution is to order to load operations correctly
so that the strategies and the autoref are the last to be loaded.
The second part is to a add a short sleep that makes sure that the
previously loaded information is propagated through all threads before
loading the strategies.