amun: trajectorypath: improve faster acceleration when only braking is necessary
Consider a robot driving the direct trajectory.
When the latter parts of the trajectory are reached, it only needs to
brake to reach its target. The problem occurs when the robot brakes
slightly slower than required by the trajectory.
For this, AlphaTimeTrajectory::findTrajectory has a mode that can
accelerate slightly faster than usually possible when generating a
trajectory.
This however only works when both the x and y components are in the
braking phase. If for example, the x component is already nearly at its
target but currently at a slow velocity, it might be required to
accelerate to a higher velocity to reach the target in a reasonable
time. This would result in the direct brake time (seen in the variable
timeDiff) of the x component to be vastly higher than the y component
even though only a low acceleration was used for it.

This commit solves the issue by allowing one of the components to
accelerate to a higher velocity in the fast brake mode if it allows for
a faster trajectory.
For the slower component, a trajectory is generated that has the time of
the faster component with an arbitrary acceleration.
If the acceleration is acceptable, this trajectory is then used.