glados: o/physics: make robotMinEndspeed results more stable
This is to fix a flickering that occurred in BreakPass.
BreakPass wants the robot to arrive before the ball and thus computes
the minEndspeed for that time.

The problem is that robotMinEndSpeed uses a binary search to find the
endspeed, but the time output of robotTimeToPos can have a jump in it.
Consider the case that robotMinEndspeed gets a target time of 0.2.
With an endspeed larger than 0.5, robotTimeToPos produces times close to
0.18 and lower for faster speeds.
With an endspeed lower than 0.5 though, robotTimeToPos produces times
larger than 0.4.
Since the exact target time can not be reached in this example,
robotMinEndspeed will return a speed around 0.5.
Which side of the jump is chosen is dependent on the exact values though
and might vary frame to frame.
Also, an endspeed lower than 0.5 might be chosen even though it is
vastly worse than an endspeed larger than 0.5 for wich the time is far
closer.

Therefore, this commit changes robotMinEndspeed such that the speed is
used that comes closest to the desired time.