ra: configdialog: enable theme/color override
Introduce a new setting enabling to choose which QStyle is used to draw
the QWidgets and additionally enables a color override to e.g. provide a
dark mode on systems that do not natively allow that (Windows for
example, at least not trivially with Qt).
Definitely not to make the UI look weird with interesting custom colors.