amun: path/speedprofile: fix usage of time for velocity/position calculation
The issue was that the parameter time (given to
partialSegmentOffsetAndSpeed) is the time in the transformed trajectory,
but the when tm is calculated, t0 was used as a base.
Since t0 was a time in the original trajectory, this can create problems
were tm is comparatively large even at the very beginning of the
segment.
This can lead to incorrect velocity, position information.