glados: Set breakpass priority lower than feintpass.
Feintpass needs to be prioritized over breakpass.
Otherwise the feinting robot would not move out of
the pass and redecide to breakpass when the pass is
shot. Since feintpass depends on the support zones,
it is only chosen in RunObjective. This means breakpass
must now also be put into the objectives.