glados: o/goal: calculate enemy shots with allowedMaxBallSpeed
Assuming that the opponent is as bad as us when it comes to precision
is not a good assumption.

Fixes #917

Most of the fix was already included in cb594bf7617ab7267d8cfed5879f6780ced4deb0:
Instead of crashing, this commits makes us print a warning (?) and handle it gracefully.

The missing part is only relevant for one of the logs: If the ball really IS faster than allowed
by the autoref, cb594bf's graceful handling is fine, but if the ball is just faster than we allow for ourselfs,
that is not a great reason to reduce the quality of predict shot.

So instead, we use the slightly faster allowed speed for our oponents