ra: plotterwidget: fix text rendering
Text rendering was broken after the change to QOpenGLWidget, because the
rendering works different now. Reordering the instructions so that the
texts get rendered after the native OpenGL calls fixes this. This may
not be the most elegant or perfomant implementation, but it does not
seem to slow down the plotter rendering and should be fine for our
purposes.
The problem was that the best way to draw a pixmap on the new
QOpenGLWidget is to use a QPainter, but the QPainter has its own context
for OpenGL, which we change when using native OpenGL calls without
calling beginNativePainting and endNativePainting. The best way to
achieve this was to buffer the QPixmaps to be rendered and render all of
them at the end, so we can wrap the native calls in the
beginNativePainting and endNativePainting calls.