glados: tr/pathhelper: prevent robots from driving behind the goal
The issue here was that during ball placement, the field border given to
the pathfinding is increased so that the robot can reach the ball even
if it is at the border.
This made it possible for a robot to try and drive behind the goal since
there is available space when the field border is not considered.

To fix this, this commit decreases the field border during ball
placement and makes the goal backside obstacle larger so that there is
no space between the goal and an enlarged field border.