data: pkg/pack-linux: Check if the build folder is built with relativeDataDirs
The question was, how do we get the information out of cmake into that script.

The first option, parsing the CMakeCache.txt feels like accessing this information
out of contract, and was disregarded.

The second option to try to pack it, de-pack it somewhere else and check (for example with amun-cli)
if our programs complain about missing references (which will only happen if the sources are temporarily moved anyways)
was disregarded as hack - and also a lot of work to work properly and without issues

Instead, a new file defs.txt.in in source was created that shall be used to store information from configure-time
as an artifact in the build folder, so the pack-linux-script can access them

However, just using {$RELATIVE_DATA_DIRS} in the defs.txt.in file is not great,
as this leaves open the problem how to figure out what cmake considers as truthy or falsy.

Instead, #cmakedefine01 is used so we can use cmake itself to tell us if
RELATIVE_DATA_DIRS has been set to a thruthy value or not.

 #cmakedefine01 will expand to #define RELATIVE_DATA_DIRS 1 if RELATIVE_DATA_DIRS is considered truthy at configure time,
and #define RELATIVE_DATA_DIRS 0 otherwise. To avoid including this defs file that is strictly just for script checking configuration options,

the script does contain an #error command to stop you, except if the very long and very obviously wrong macro
'CHEAT_DEFS_H_ERROR_ABORTION_DO_NOT_USE_IN_CODE_OR_WE_WILL_HATE_YOU' is defined.

To use this defs.txt file generated at configure time, we feed RELATIVE_DATA_DIRS
into gcc, with necessary parameters for gcc to 1. accept inputs from stdin, 2. include defs.txt before any contents from stdin
and 3. get around the #error directive in src/defs.txt