amun: trajectorypath: fix calculateTrajectory when start and end speed are the same
When the start and end speed are close to each other, a simple
trajectory that did basically nothing was returned, to avoid numeric
issues later in the calculation. The trajectory set the
speed to zero and was only fractions of a millisecond long.

This case is only valid when the time is also close to zero, as then we
actually dont want to move the robot a lot (or at all). When the time is
non zero, then we want to return the calculated trajectory.