amun: trajectorypath: fix acceleration calculation in slowdown acceleration profiles
The problem with the acceleration calculation was that it was computed
as if the acceleration change is linear in the untransformed trajectory
time.
The issue is that the acceleration is only linear in the transformed
trajectory. In the untransformed trajectory, the acceleration is
nonlinear.

The acceleration is only dependent on the time to the end of the
trajectory and does not actually depend on the current velocity of the
segment. This was only included for simplicity.
The previous code could be refactored to not use the velocity for the
time computation.

The new function computeAcceleration is the corrected version of the
acceleration computation. It computes the inverse of the integral of the
acceleration in the transformed trajectory and thereby finds the correct
acceleration factor.