ra: fieldwidget: fix render scaling of SVG screenshotting
Previously, the scaling depended on the current scaling of the
fieldwidget in ra, which can create problems when using the results.
A factor of 100 is used since the image is only 12 "pixels" wide
otherwise.
While this should not be an issue for vector graphics in theory, some
PDF viewers can not correctly render the result (probably due to
floating point/scaling issues).