glados: t/a/acceptpass Offset position according to attackpos instead of ball position
Especially during feintpass we can run into the situation that
the current ball position is not in the same direction from which
we are going to receive the pass.
Example Setup: Robot A plans to pass to Robot B. Robot C does
feintpass and positions itself between B and A. Robot A shoots the
Ball and B becomes MA. B immediately decides to pass the Ball to C.
Now the ball is between A and C but C will actually recieve the
pass from B.

Therefore the timing considerations from C are actually off,
since it will need to reposition and turn around before it can
recieve the pass. By using the attackposition instead of the
ball position (if it exists) we will automatically consider
the direction from which the ball will be coming to C.