glados: u/attack Rate passes badly where the ball needs to go through the accepting robot
Consider the following situation that arises from feintpass:
Robot A plans a pass to robot B. Robot C does feintpass and
positions itself between A and B. A shoots the ball and B
becomes MA. Now B plans to pass the ball to C. However, C
has to evade the incoming ball first, invalidating the pass
timing of C.

The 'correct' way to deal with this is probably to adjust Cs
timing beforehand. However, this could be rather unstable,
since we would need determine if C reaches the passline before
or after the ball. This consideration is potentially subject
to a lot of noise.

Therefore we just rate passes badly, if two conditions are met:
1. The angle pass start - World.Ball.pos - pass end is small,
i.e. the ball needs to travel through one of the robots
2. The pass start position is further from the ball than the pass
destination, i.e. the ball needs to travel through the accepting
robot (Passes shot in the same direction as the incoming ball, e.g.
by stopping it first and rotating around it should not be penalized)