amun: processor/radiosystem: Ask transceivers for their status prior to construction
When supporting multiple transceiver there is a large number of possible
configurations: each single transceiver may be unsupported (for example
because the binary was built without support for libusb), unplugged and
plugged in. If there are n different supported transceivers, there are
3^n such combinations.

Two such configurations should be disallowed
1 Multiple transceivers are plugged in and supported. In such cases, it
  is not immediately clear to the user which transceiver is currently
  used.
2 No transceiver is plugged in and supported.

Configuration 1 is not totally non sensical. We may want to address both
old and new robots at the same time. However implementing this is more
complex and not done for now.

To forego constructing each transceiver and then checking wether the
configuration is valid, their status is checked beforehand. To avoid the
combinatorial explosion, the status is stored in an array.