amun: gamecontroller: Rename classes
In the past, the name InternalGameController corresponded to the
reimplementation of the whole game controller logic in C++ while
ExternalGameController was a game controller reachable through a network
connection. After the official game controller binary was integrated,
the InternalGameController was deleted and the SSLGameController class
was added, used to talk to a game controller process started by Ra.
Since the CI protocol uses TCP, the ExternalGameController was reused.
Since these names can be confusing (e.g. ExternalGameController is used
by SSLGameController), they were renamed.

GameControllerConnection is renamed to StrategyGameControllerMediator
since it multiplexes a strategy scripts (team or autoref) connection to
the game controller (i.e. to the internal or an external game
controller)

SSLGameController is renamed to InternalGameController since both an
externally running and internally running game controller are a SSL game
controller. Internal should clarify that this is the game controller
started by Ra.

ExternalGameController is renamed to GameControllerSocket as it holds a
TCP socket and is used both by the StrategyGameControllerMediator and by
the InternalGameController