glados: u/attack: Don't duplicate the PassInfo type
The target field can be made optional since users of passInfo do not
rely on the pass being non anonymous. This allows the removal of one
type assertion in a/a/freekick.

SideStep could use Pick<PassInfo, "ballPos" | "time"> for more
generality, however this was not done to allow easier usage of the other
PassInfo fields later on (i.e. be more flexible)