glados: a/a/freekick: Don't send old passInfo when time is running out
Whenever _passList is set, it is sent out. Even when the state is
ShootGoal, _passList may have been set in an earlier iteration: a pass
may have been selected, then the time started running out, we switch to
Wait and then to ShootGoal.

However, when the state is changed to ShootGoal, the pass will never be
played as FreeKick will remain in ShootGoal and possible receivers are
better off to to something else. It is thus not sensible to send out the
message in this case.