glados: a/a/freekick: Use tagged enum for state
This encodes the fact that _pass and _passList are always set
together/only used in PassPrepare/Pass in the type system.

sortPassesFromSuggestions is only used in these two instances and need
not return undefined when no suitable pass is found since no local state
is overwritten in this case now.